Fix device_id test case
DeviceID property of port is determined when some resource
is connected to the port.

But in current device id test, nothing will be attached,
so in all test execution, device_id is actually set as blank string.

This means that if some other port exists in same tenant,
all of test execution turns to be failed.

So I fixed test case as followings.

- Create subnet in test
- Create instance with fixed ip

Then port.device_id will be set with the ID of instance,
so test will be work expectedly.